+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.3
+* Set min SDK version to `>=2.0.0`.
+* Set max analyzer version to `<0.34.0`
+* Support ContentType constants.
+* Support InternetAddress constants.
+* Support HttpHeaders constants.
+* Update travis to use the latest stable version of Dart.
+
 ## 1.0.2
 
 * Widen SDK constraint to include 2.0.0.